Relabel tests as unit type.
